backend/requirements.txt
fastapi
uvicorn
websockets
spotipy
yt-dlp
python-dotenv